Fix lcov and change regression node label
The line coverage tests ran smoke before regression  Smoke, after it's
run cleans up the /var/lib/glusterd directory. This directory is
required for a bunch of different tests. So if we try to run regressions
without it, we create unexpected results.

It's time to move all of regressions to exclusively run on Centos 7. All
Centos 6 machines are officially now deprecated. We'll use a new tag and
add more Centos 6 machines as the week goes.

Change-Id: Idd38dd83c8c484163d8c62879868d318517eedab