Automatically accept the review.gluster.org ssh key
See https://bugzilla.redhat.com/show_bug.cgi?id=1698694

While this would open us to MITM in very specific case (like, just
after a server was reinstalled), a attacker would like just get
access to vote on a patch as the regression tests, which do not seems
that a big attack. This wouldn't also be stealth, since the MITM would
be discovered as soon as the traffic is no longer intercepted.

Plus, the others scripts already do the same.

Change-Id: I8e294d92c2602c86fe6f04227512ce70ec081f03