Adding the generic_package.sh script file
This script file is used to build the rpms for
Debian and Ubuntu versions.

We can run this script on a ubuntu machine.
We create a chroot environment using pbuilder create.
This gives us the environment for both ubuntu and debian.
On this we build, sign and push the packages.

We update the changelogs and push them to the glusterfs-debian
repo as well.

Signed-off-by: Sheetal Pamecha <spamecha@redhat.com>
Signed-off-by: hari gowtham <hgowtham@redhat.com>
Signed-off-by: Shwetha K Acharya <sacharya@redhat.com>